fix: update rq worker dependency setup
