+++ src/infrastructure/rq/tasks.py
@@ -1,16 +1,16 @@
+from __future__ import annotations
+
 import asyncio
 
-from src.infrastructure.common.logger import configure_logging, get_logger
+from src.infrastructure.common.logger import get_logger
+from src.infrastructure.rq.rq_worker import services
+from src.infrastructure.rq.worker_services import WorkerServices
 
-
-configure_logging(debug=True)
 logger = get_logger()
 
 
-def process_documents(job_id: str, documents: str) -> None:
+def process_documents(job_id: str, documents: str, services: WorkerServices = services) -> None:
     """Load documents into the vector database."""
-    from src.infrastructure.rq.rq_worker import services
-
     logger.info(f"Processing documents for job: {job_id}")
     chroma_client = services.chroma_client
     asyncio.run(chroma_client.client.heartbeat())
